system/requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
attrs==23.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
cachetools==5.3.2
cchardet==2.1.7
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
courlan==0.9.5
cssselect==1.2.0
datasets==2.16.1
dateparser==1.2.0
dill==0.3.7
distro==1.9.0
etils==0.9.0
exceptiongroup==1.2.0
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.13.1
frozenlist==1.4.1
fsspec==2023.10.0
google-ai-generativelanguage==0.1.0
google-api-core==2.15.0
google-auth==2.26.2
google-generativeai==0.1.0rc1
googleapis-common-protos==1.57.0
grpcio==1.60.0
grpcio-status==1.60.0
h11==0.14.0
htmldate==1.6.1
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.20.2
idna==3.6
importlib-resources==5.10.2
jieba3k==0.35.1
joblib==1.3.2
jusText==3.0.0
keras==2.11.0
langcodes==3.3.0
lxml==4.9.4
multidict==6.0.4
multiprocess==0.70.15
newspaper3k==0.2.8
nltk==3.8.1
numpy==1.24.4
openai==1.6.1
packaging==23.2
pandas==2.0.3
pillow==10.2.0
promise==2.3
proto-plus==1.23.0
protobuf==4.25.2
pyarrow==14.0.2
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.5.3
pydantic_core==2.14.6
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
rank-bm25==0.2.2
regex==2023.12.25
requests==2.31.0
requests-file==1.5.1
rsa==4.9
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
tensorflow==2.11.0
tensorflow-datasets==4.8.0
tensorflow-intel==2.11.0
tensorflow-metadata==1.12.0
tinysegmenter==0.3
tld==0.13
tldextract==5.1.1
tqdm==4.66.1
trafilatura==1.6.3
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
urllib3==2.1.0
xxhash==3.4.1
yarl==1.9.4
